Corrections from review
