+++ sustainml_cpp/include/sustainml_cpp/nodes/AppRequirementsNode.hpp
@@ -95,8 +95,8 @@ public:
 
     SUSTAINML_CPP_DLL_API AppRequirementsNode(
             AppRequirementsTaskListener& user_listener,
-            sustainml::core::Options opts,
-            sustainml::core::RequestReplyListener& req_res_listener);
+            sustainml::core::RequestReplyListener& req_res_listener,
+            sustainml::core::Options opts);
 #endif // SWIG_WRAPPER
 
     SUSTAINML_CPP_DLL_API virtual ~AppRequirementsNode();
