+++ sustainml_cpp/include/sustainml_cpp/nodes/CarbonFootprintNode.hpp
@@ -102,8 +102,8 @@ public:
 
     SUSTAINML_CPP_DLL_API CarbonFootprintNode(
             CarbonFootprintTaskListener& listener,
-            sustainml::core::Options opts,
-            sustainml::core::RequestReplyListener& req_res_listener);
+            sustainml::core::RequestReplyListener& req_res_listener,
+            sustainml::core::Options opts);
 #endif // SWIG_WRAPPER
     SUSTAINML_CPP_DLL_API virtual ~CarbonFootprintNode();
