+++ sustainml_cpp/include/sustainml_cpp/nodes/MLModelMetadataNode.hpp
@@ -95,8 +95,8 @@ public:
 
     SUSTAINML_CPP_DLL_API MLModelMetadataNode(
             MLModelMetadataTaskListener& user_listener,
-            sustainml::core::Options opts,
-            sustainml::core::RequestReplyListener& req_res_listener);
+            sustainml::core::RequestReplyListener& req_res_listener,
+            sustainml::core::Options opts);
 #endif // SWIG_WRAPPER
 
     SUSTAINML_CPP_DLL_API virtual ~MLModelMetadataNode();
