+++ sustainml_cpp/include/sustainml_cpp/nodes/MLModelNode.hpp
@@ -105,8 +105,8 @@ public:
 
     SUSTAINML_CPP_DLL_API MLModelNode(
             MLModelTaskListener& user_listener,
-            sustainml::core::Options opts,
-            sustainml::core::RequestReplyListener& req_res_listener);
+            sustainml::core::RequestReplyListener& req_res_listener,
+            sustainml::core::Options opts);
 #endif // SWIG_WRAPPER
 
     SUSTAINML_CPP_DLL_API virtual ~MLModelNode();
