+++ sustainml_cpp/src/cpp/nodes/AppRequirementsNode.cpp
@@ -47,6 +47,7 @@ AppRequirementsNode::AppRequirementsNode(
     opts.wqos.resource_limits().max_instances = 500;
     opts.wqos.resource_limits().max_samples_per_instance = 1;
     opts.wqos.durability().kind = eprosima::fastdds::dds::TRANSIENT_LOCAL_DURABILITY_QOS;
+
     init(opts);
 }
 
@@ -81,8 +82,8 @@ AppRequirementsNode::AppRequirementsNode(
 
 AppRequirementsNode::AppRequirementsNode(
         AppRequirementsTaskListener& user_listener,
-        sustainml::core::Options opts,
-        sustainml::core::RequestReplyListener& req_res_listener)
+        sustainml::core::RequestReplyListener& req_res_listener,
+        sustainml::core::Options opts)
     : Node(common::APP_REQUIREMENTS_NODE, opts, req_res_listener)
     , user_listener_(user_listener)
 {
