+++ sustainml_cpp/src/cpp/nodes/MLModelMetadataNode.cpp
@@ -46,6 +46,7 @@ MLModelMetadataNode::MLModelMetadataNode(
     opts.wqos.resource_limits().max_instances = 500;
     opts.wqos.resource_limits().max_samples_per_instance = 1;
     opts.wqos.durability().kind = eprosima::fastdds::dds::TRANSIENT_LOCAL_DURABILITY_QOS;
+
     init(opts);
 }
 
@@ -66,6 +67,7 @@ MLModelMetadataNode::MLModelMetadataNode(
     opts.wqos.resource_limits().max_instances = 500;
     opts.wqos.resource_limits().max_samples_per_instance = 1;
     opts.wqos.durability().kind = eprosima::fastdds::dds::TRANSIENT_LOCAL_DURABILITY_QOS;
+
     init(opts);
 }
 
@@ -80,8 +82,8 @@ MLModelMetadataNode::MLModelMetadataNode(
 
 MLModelMetadataNode::MLModelMetadataNode(
         MLModelMetadataTaskListener& user_listener,
-        sustainml::core::Options opts,
-        sustainml::core::RequestReplyListener& req_res_listener)
+        sustainml::core::RequestReplyListener& req_res_listener,
+        sustainml::core::Options opts)
     : Node(common::ML_MODEL_METADATA_NODE, opts, req_res_listener)
     , user_listener_(user_listener)
 {
